Munene-Sylvia-Backend/requirements.txt
Django==6.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-spectacular==0.27.0
python-decouple==3.8
requests==2.31.0
Pillow>=12.0.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
PyMySQL==1.1.2
firebase-admin
